Modifications done by Oussama Kanoun and commited by Florent Lamiraux
In jrlHumanoidDynamicRobot.h, added a list of fixed joints : joints that
are temporarily in contact with a fixed part of the environment (ground for
instance).
In jrlJoint.h, added a method to compute the Jacobian of the joint wrt the
internal configuration variables, a method to get the result of this
computation and a method to compute the Jacobian of a point of the joint.